llm-based-assistants/infos/llm_inference_guide/requirements.txt
langchain_nvidia_ai_endpoints==0.3.9
python-dotenv==1.1.0
ratelimiter==2.2.1